+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/DelegationRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.Delegation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DelegationRepository extends JpaRepository<Delegation, Long> {
+public interface DelegationRepository extends JpaRepository<Delegation, Integer> {
 }
