+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/InfrastructureSportiveRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.InfrastructureSportive;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InfrastructureSportiveRepository extends JpaRepository<InfrastructureSportive, Long> {
+public interface InfrastructureSportiveRepository extends JpaRepository<InfrastructureSportive, Integer> {
 }
