+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/ResultatRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.Resultat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResultatRepository extends JpaRepository<Resultat, Long> {
+public interface ResultatRepository extends JpaRepository<Resultat, Integer> {
 }
