+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/SpectateurRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.Spectateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpectateurRepository extends JpaRepository<Spectateur, Long> {
+public interface SpectateurRepository extends JpaRepository<Spectateur, Integer> {
 }
